+++ nano/node/fwd.hpp
@@ -6,9 +6,14 @@
 
 namespace nano
 {
+class active_transactions;
 class ledger;
+class local_vote_history;
 class logger;
-class node;
 class network;
+class node;
+class node_config;
 class stats;
+class vote_generator;
+class wallets;
 }
